@@ -31,7 +31,7 @@ enum CMExtensionName: String {
 }
 
 /// Represents a cmark extension option.
-public struct CMExtensionOption: OptionSet {
+public struct CMExtensionOption: OptionSet, Sendable {
 
     /// The raw value.
     public let rawValue: Int32
